readyup|refactor: Extract the run command's staleness advisories
Extracts `readManifestTracking` and `warnOnKitStaleness` with their two helpers out of `readyup`'s `src/run/runCommand.ts` into `kit-staleness.ts`, which takes a kebab-case name for having no single main export. `runCommand.ts` reaches neither the manifest reader nor the drift checks now, and holds no execution-kernel declaration at all: what remains is the two output modes and their rendering.

`kit-staleness.unit.test.ts` retires `stalenessWarning.unit.test.ts`, splitting it along the two altitudes it mixed. `warnOnKitStaleness` is now driven against a `ManifestTracking` value built in the test rather than one read through a mocked manifest, so every advisory and every silence is stated against the input that decides it.

`runCommand.unit.test.ts` keeps what the run itself decides -- one manifest read per invocation however many kits follow, the `--jit` flag reaching that read, one advisory pass per kit against the source it resolved to, the raised advisories arriving in the JSON report, the `warnings` field absent when none were raised, and the exit code untouched throughout -- and mocks `kit-staleness.ts` directly, that seam being what the run consumes.
