+++ packages/readyup/src/run/runCommand.ts
@@ -7,18 +7,14 @@ import type { KitProvenance } from '../kits/KitProvenance.ts';
 import type { FixLocation, RdyChecklist, RdyKit, RdyReport, RdyStagedChecklist, Severity } from '../kits/types.ts';
 import { getLayout } from '../layout/engine.ts';
 import type { BreadcrumbSegment, SummaryRow } from '../layout/layoutEngine.ts';
-import { DEFAULT_MANIFEST_PATH } from '../manifest/manifestPath.ts';
-import type { RdyManifest } from '../manifest/manifestSchema.ts';
-import { readManifest } from '../manifest/readManifest.ts';
 import { formatCombinedSummary } from '../reporting/formatCombinedSummary.ts';
 import { formatJsonReport, type KitInput } from '../reporting/formatJsonReport.ts';
 import { countResults, reportRdy } from '../reporting/reportRdy.ts';
-import type { JsonWarning, RaisedWarning } from '../schemas/common.ts';
+import type { JsonWarning } from '../schemas/common.ts';
 import type { JsonDetail, JsonKitOrigin } from '../schemas/reportSchema.ts';
-import { checkDrift } from '../verify/checkDrift.ts';
-import { checkSourceDrift } from '../verify/checkSourceDrift.ts';
+import { readManifestTracking, warnOnKitStaleness } from './kit-staleness.ts';
 import { loadKit } from './loadKit.ts';
-import type { KitSource, ResolvedKitEntry } from './ResolvedKitEntry.ts';
+import type { ResolvedKitEntry } from './ResolvedKitEntry.ts';
 import { resolveRunExitCode } from './resolveRunExitCode.ts';
 import { resolveThresholds } from './resolveThresholds.ts';
 import { runRdy } from './runRdy.ts';
@@ -47,100 +43,6 @@ interface HumanRunSettings {
   failOn: Severity | undefined;
   quiet: boolean;
   reportOn: Severity | undefined;
-}
-
-/** The manifest an invocation checks its kits against, read once and shared by every kit in the run. */
-interface ManifestTracking {
-  manifest: RdyManifest;
-  manifestDir: string;
-}
-
-/**
- * Read the default manifest for the run's advisories, best effort.
- *
- * Every failure here answers with no manifest at all: a missing one is the normal state of a
- * project that never compiled, and an unreadable or unrecognized one says nothing about any kit. A
- * verification tool that refused to run because its own bookkeeping was unreadable would be worse
- * than one that runs and stays quiet. `--jit` runs from source, which the manifest does not
- * describe, so they skip the read entirely.
- */
-function readManifestTracking(isJit: boolean): ManifestTracking | undefined {
-  if (isJit) return undefined;
-  const manifestPath = path.resolve(process.cwd(), DEFAULT_MANIFEST_PATH);
-  try {
-    return { manifest: readManifest(manifestPath), manifestDir: path.dirname(manifestPath) };
-  } catch {
-    return undefined;
-  }
-}
-
-/**
- * Emit advisory stderr warnings when the manifest disagrees with the kit that is about to run.
- *
- * `target-drift` says the compiled bundle is not the one the manifest recorded, so someone edited
- * it by hand. `source-stale` says the TypeScript it was built from has moved on, so the run is
- * about to execute checks that no longer match their source. Both can hold at once.
- *
- * Advisory by design: `rdy verify` is the enforcing gate, and this never touches the exit code. A
- * kit no entry describes, an entry recording no hash, a remote or just-in-time source, and a file
- * that cannot be hashed are all silent, because none of them is evidence that anything is stale.
- *
- * The stderr lines are written in both modes; the returned entries are what JSON mode captures into
- * the report, so a consumer that owns only stdout still learns the run was advised of something.
- */
-function warnOnKitStaleness(
-  kitName: string,
-  source: KitSource,
-  tracking: ManifestTracking | undefined,
-): RaisedWarning[] {
-  if (tracking === undefined || 'url' in source) return [];
-
-  const entry = findManifestEntry(source.path, tracking);
-  if (entry === undefined) return [];
-
-  const warnings: RaisedWarning[] = [];
-  if (hasVerdict(() => checkDrift(entry, tracking.manifestDir), 'drift')) {
-    warnings.push({
-      code: 'target-drift',
-      message: `compiled kit "${kitName}" does not match the hash the manifest recorded for it.`,
-      remedy: 'Run `rdy compile --force` to rebuild it from source.',
-    });
-  }
-  if (hasVerdict(() => checkSourceDrift(entry, tracking.manifestDir), 'stale')) {
-    warnings.push({
-      code: 'source-stale',
-      message: `kit "${kitName}" was compiled from an older source than the one on disk.`,
-      remedy: 'Run `rdy compile` to rebuild it.',
-    });
-  }
-
-  for (const warning of warnings) {
-    process.stderr.write(`Warning: ${warning.message} ${warning.remedy}\n`);
-  }
-  return warnings;
-}
-
-/**
- * Find the manifest entry describing a kit, matching on resolved compiled path.
- *
- * Matching by name instead would misfire wherever a kit's name and its file part company: `--file`
- * names a kit by an arbitrary path, and a custom `outDir` puts a differently-named entry's output
- * where this one's would go.
- */
-function findManifestEntry(kitPath: string, tracking: ManifestTracking): RdyManifest['kits'][number] | undefined {
-  const resolvedKitPath = path.resolve(process.cwd(), kitPath);
-  return tracking.manifest.kits.find(
-    (kit) => kit.path !== undefined && path.resolve(tracking.manifestDir, kit.path) === resolvedKitPath,
-  );
-}
-
-/** Report whether a staleness predicate reaches the given verdict, treating a file it cannot hash as no. */
-function hasVerdict<TStatus extends { kind: string }>(check: () => TStatus, kind: TStatus['kind']): boolean {
-  try {
-    return check().kind === kind;
-  } catch {
-    return false;
-  }
 }
 
 /** Run rdy checklists across one or more kits. Returns a numeric exit code. */
